fix: pomodoro and todolist working for unaccount

+++ client/stores/pomodoros.ts
@@ -1,9 +1,12 @@
 import { computed, ref } from 'vue'
 import { useRuntimeConfig } from '#app'
 import { StorageSerializers, createGlobalState, useStorage } from '@vueuse/core'
+import {useAuthStore} from "./auth"
+
 
 export const usePomodoroStore = createGlobalState( () => {
   const API_URL = useRuntimeConfig().public.API_URL
+  const { isAuth } = useAuthStore()
 
   // state
   const currentPomodoroSection = ref()
@@ -12,8 +15,9 @@ export const usePomodoroStore = createGlobalState( () => {
 
   // actions
   const startPomodoro = async() =>  {
-        const response = await fetchWithAuth(`${API_URL}/pomodoros/`,{
-            method: "POST",
+    if(!isAuth.value) return
+        const response = await fetchWithAuth(`${API_URL}/pomodoros/`, {
+          method: "POST",
         });
         if (response.ok){
             currentPomodoroSection.value = await response.json()
@@ -22,14 +26,16 @@ export const usePomodoroStore = createGlobalState( () => {
     }
 
   const pomodoroSectionAction = async(action: string) =>  {
+    if(!isAuth.value) return
         const response = await fetchWithAuth(`${API_URL}/pomodoros/${currentPomodoroSection.value.id}`,{
-            method: "PATCH",
-			body: JSON.stringify({action: action}),
+          method: "PATCH",
+			    body: JSON.stringify({action: action}),
         });
         if (!response.ok) throw new Error(`Fail to ${action} pomodoro section`);
     }
 
   const deletePomodoro = async() =>  {
+    if(!isAuth.value) return
         const response = await fetchWithAuth(
           `${API_URL}/pomodoros/_last`,
           {method: "DELETE"}
@@ -48,6 +54,6 @@ export const usePomodoroStore = createGlobalState( () => {
     // actions
     startPomodoro,
     pomodoroSectionAction,
-deletePomodoro,
+    deletePomodoro,
   }
 })
