+++ client/stores/todolist.ts
@@ -1,7 +1,8 @@
 import { ref } from 'vue'
 import { useRuntimeConfig } from '#app'
-import { createGlobalState } from '@vueuse/core'
+import { createGlobalState, useLocalStorage } from '@vueuse/core'
 import _ from "lodash";
+import {useAuthStore} from "./auth"
 
 export enum TaskStatus {
   TODO = "TO-DO",
@@ -12,25 +13,35 @@ export enum TaskStatus {
 
 export const useTaskStore = createGlobalState( () => {
   const API_URL = useRuntimeConfig().public.API_URL
+  const { isAuth } = useAuthStore()
 
   // state
-  const tasks = ref([])
-  const enableTodoListTask = ref(true)
+  const tasks = useLocalStorage('tasks', [])
+  const enableTodoListTask = useLocalStorage('enableTodoListTask', true)
 
   // getters
 
   // actions
   const getTaskList = async() =>  {
+    if(!isAuth.value) return
         const response = await fetchWithAuth(`${API_URL}/todolist`);
         if (response.ok) tasks.value = await response.json();
         else throw new Error('abc?');
         tasks.value = tasks.value.map((task)=>({
           ...task,
-          "section": "work",
+          section: "work",
         }))
     }
 
   const postTask = async (title: string) =>  {
+    if(!isAuth.value) {
+      tasks.value.push({
+        title,
+        status: TaskStatus.DOING,
+        section: "work",
+      })
+      return
+    }
     const index = _.isEmpty(tasks.value)
     ? 1
     : Math.max(...tasks.value.map(task => task.index)) + 1
@@ -44,6 +55,7 @@ export const useTaskStore = createGlobalState( () => {
     }
 
   const patchTask = async (taskId: string, change = {}) =>  {
+    if(!isAuth.value) return
     if(_isEmpty(change)) return
     const response = await fetchWithAuth(`${API_URL}/todolist/${taskId}`,
     {
@@ -53,6 +65,7 @@ export const useTaskStore = createGlobalState( () => {
     }
 
   const deleteTask = async (taskId: string) =>  {
+    if(!isAuth.value) return
       const response = await fetchWithAuth(`${API_URL}/todolist/${taskId}`, {method: "DELETE"})
       if (response.ok) tasks.value = tasks.value.filter(task => task.id !== taskId)
       else throw new Error('abc?');
@@ -62,7 +75,6 @@ export const useTaskStore = createGlobalState( () => {
   const moveTask = (task, newIndex: number) => {
       const oldIndex = tasks.value.indexOf(task)
       if (oldIndex < 0 || newIndex >= tasks.value.length) return
-      console.log(oldIndex, newIndex, tasks.value[oldIndex], tasks.value[newIndex], tasks.value[oldIndex].index, tasks.value[newIndex].index )
 
       const swapProp = (obj1, obj2, prop) => {
         if (!(prop in obj1) || !(prop in obj2)) return;
